backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.4
python-dotenv==1.0.1
requests==2.32.3
django-cors-headers==4.3.1
Pillow==10.3.0
